==> notebooks/2_limpar_dados.ipynb <==
## Limpar dados

Esse notebook está responsável por limpar os dados baixados. Para isso, são feitas algumas investigações para entender melhor os dados e ajudar a definir qual abordagem de limpeza utilizar para cada caso.

---

# importando pacotes
import os
import pandas as pd
import numpy as np

# agora que temos os dados armazenados, podemos criar o nosso dataframe e seguir com as análises e limpeza
pasta_atual = os.getcwd()
pasta_pai = os.path.dirname(pasta_atual)
pasta_data = os.path.join(pasta_pai, "data")

df = pd.read_csv(f"{pasta_data}\\dados_brutos.csv", parse_dates=["date"])\
    .sort_values(by=["date", "appId"])\
    .reset_index(drop=True)

df

# Vamos começar olhando uma visão geral sobre os valores nulos do dataframe
df.isna().sum()

# Valores de appId nulos não nos interessam, então vamos começar descartando eles:
df = df.loc[~df["appId"].isna()]

df.isna().sum()

# Vamos ver agora se todos os valores restantes apresentam o mesmo padrão
regex = r"^com\.app\.\d+$"
df.loc[~df["appId"].str.match(regex, na=False), "appId"]

# Todos os valores de appId apresentam o mesmo padrão e não apresentam valores nulos, podemos seguir em frente.

# Não existem valores nulo de data, mas isso não significa que não podem haver valores inválidos
sorted(df["date"].dt.year.unique())

# Todos os anos exceto 2024 são anos inválidos e podem ser filtrados
df = df.loc[df["date"].dt.year == 2024]

df.isna().sum()

# ver se existem combinações repetidas de appId e date
df[df.duplicated(subset=["appId", "date"])]

# deletar duplicatas
df = df.drop_duplicates()

# vamos ver se ainda existem combinações repetidas de appId e date depois de deletar linhas duplicadas
df[df.duplicated(subset=["appId", "date"])]

# como não temos casos de dados de um appId mais de uma vez no mesmo dia, para auxiliar nas análises futuras, vamos garantir
# que para cada dia temos todas as combinações de appId e data
quantidade_original = df.shape[0]

periodo_completo = pd.date_range(start=df["date"].min(), end=df["date"].max())

app_ids = df["appId"].unique()

todas_combinacoes = pd.MultiIndex.from_product([app_ids, periodo_completo], names=["appId", "date"])

df = df.set_index(["appId", "date"]).reindex(todas_combinacoes).reset_index()
print(f"Dados adicionados: {df.shape[0] - quantidade_original}")

# Vamos seguir olhando as columnas "country" e "lang"
print(f"valores distintos 'country': {df["country"].nunique()}")
print(f"valores distintos 'lang': {df["lang"].nunique()}")

# ambas as colunas apresentam apenas um único valor, portanto essas colunas não agregam valor e podem ser descartadas
df = df.drop(["country", "lang"], axis=1)

# Vamos entender a coluna "category".
# Primeiro, vamos ver se existem casos em que um aplicativo pode ter mais de uma categoria
df.groupby("appId")["category"]\
    .nunique()\
    .reset_index(name = "contagem_categorias")\
    .query("contagem_categorias >= 2")\
    .sort_values(by = "contagem_categorias", ascending = False)

# Não existem casos de um aplicativo ter mais de uma categoria, mas vamos ver se existem aplicativos com
# uma categoria definida e também com valores nulos
df.groupby("appId")["category"]\
    .nunique(dropna=False)\
    .reset_index(name = "contagem_categorias")\
    .query("contagem_categorias >= 2")\
    .sort_values(by = "contagem_categorias", ascending = False)

# Esse valor foi populado pela tabela "ratings_reviews", e provavelmente se refere a categoria do review.
# Entretanto, como observamos apenas valores válidos exclusivos por aplicativo, podemos atribuir a categoria
# ao aplicativo independente de ter ocorrido uma avaliação.
mapa = df.dropna(subset=["category"]).drop_duplicates("appId").set_index("appId")["category"].to_dict()

df["category"] = df.apply(
    lambda linha: mapa.get(linha["appId"], "NO_CATEGORY") if pd.isna(linha["category"]) else linha["category"],
    axis=1
)

df.isna().sum()

# não existem mais valores nulos para "Category", vamos validar se existem casos de apps com mais de uma categoria
df.groupby("appId")["category"]\
    .nunique()\
    .reset_index(name = "contagem_categorias")\
    .query("contagem_categorias >= 2")\
    .sort_values(by = "contagem_categorias", ascending = False)

# as colunas "ratings" e "reviews" são cumulativas e, portanto, nunca devem apresentar valor nulo (exceto
# a primeira ocorrencia).
# A princípio, vamos preencher esses valores, mas precisamos tomar cuidado pois também existem as colunas "daily_ratings"
# e "daily_reviews" que são relacionadas.
df.loc[
    (df["ratings"].isna() ^ df["daily_ratings"].isna())
    | (df["reviews"].isna() ^ df["daily_reviews"].isna())
]

# não existem casos em que o valor diário está nulo mas o cumulativo está definido, e vice versa.

# quando existir valores nulos de "ratings" e "reviews", vamos atribuir que não teve novas avaliações nesse período,
# e preencher com o ultimo valor observado
df[["ratings", "reviews"]] = df.groupby("appId")[["ratings", "reviews"]]\
      .transform(lambda x: x.ffill())

df.isna().sum()

# Ainda restam os casos em que "reviews" e "ratings" começam com valor nulo (sem dados para serem propagados)
# Entretanto, não sabemos como era o comportamento dessas avaliações antes do primeiro dado armazenado.
# Vamos adotar que antes da primeira avaliação, não haviam avaliações
df[["ratings", "reviews"]] = df[["ratings", "reviews"]].fillna(0)

df.isna().sum()

# vamos recalcular as colunas de "daily_ratings" e "daily_reviews", presumindo que o banco de dados é atualizado consultando 
# a quantidade de avaliações diárias, pois é uma informação mais garantida que apenas olhar para a quantiadde de novas
# avaliações 
df["daily_ratings"] = df.groupby("appId")["ratings"].shift(-1) - df["ratings"] 
df["daily_reviews"] = df.groupby("appId")["reviews"].shift(-1) - df["reviews"] 


df.isna().sum()

# ainda sobram os casos em que daily_ratings e daily_reviews são nulos, pois não temos o valor das avaliações seguintes.
# para esses casos, vamos replicar o último valor observado
df[["daily_ratings", "daily_reviews"]] = df.groupby("appId")[["daily_ratings", "daily_reviews"]]\
      .transform(lambda x: x.ffill())

df.isna().sum()

# para os valores restantes, existem alguns casos que podem explicar valores nulos, vamos adotar a seguinte regra:
# para mauReal e predictionLoss, vamos interpolar os dados possíveis, pois é mais improvável que tenhamos 30 dias
# sem usuários, e um período sem nenhuma desinstalação
df[["mauReal", "predictionLoss"]] = df[["mauReal", "predictionLoss"]].interpolate(method="linear")

df.isna().sum()

# não existem mais dados nulos de predictionLoss, mas ainda existem dados nulos de mauReal
# Replicar os dados finais para mauReal
df["mauReal"] = df.groupby("appId")["mauReal"]\
    .transform(lambda x: x.ffill())

df.isna().sum()

# dados finais não foram replicados, vamos investigar os dados que mauReal é zero
df.loc[df["mauReal"].isna(), "appId"].unique()

# esse erro acontece apenas com um único aplicativo
df.loc[df["appId"] == "com.app.10626"]

# esse aplicativo apresenta um valor alto de usuários desde o inicio.
# como encontramos valores de rating e reviews desde o início, vamos presumir que o app já havia sido lançado antes
# de 2024-03-23
df["mauReal"] = df.groupby("appId")["mauReal"]\
    .transform(lambda x: x.bfill())

df.isna().sum()

# Não sabemos como o banco de dados define a informação e "newinstalls", entretanto, é um tipo de informação que esperamos
# ter mais formas de conseguir essa informação correta. Ou seja, caso ela não exista, vamos presumir que ela não exista
# pois não ocorreram novas isntalações
df["newinstalls"] = df["newinstalls"].fillna(0)

df.isna().sum()

# Por fim, vamos analisar DAU.
# Não sabemos como essa informação é calculada, mas diferente de "newisntalls", existem mais situações em que podem ocorrer
# erros.
# Vamos então interpolar os dados sempre que mauReal for maior que zero
filtro = df["mauReal"] > 0
df.loc[filtro, "dauReal"] = df["dauReal"].where(filtro).interpolate()

df.isna().sum()

# quase todos os valores foram corrigidos, vamos observar os valores restantes
df.loc[df["dauReal"].isna(), "appId"].unique()

# localizamos o mesmo caso anterior
df.loc[df["appId"] == "com.app.10626"]

# nesse caso, o aplicativo apresenta ter muitos usuários mesmo antes do dia 2024-03-23. Vamos então preencher esses dados
# com o primeiro dado disponível
df["dauReal"] = df.groupby("appId")["dauReal"]\
    .transform(lambda x: x.bfill())

df.isna().sum()

# agora que não existem mais nenhum valor nulo, vamos ver se todos os valores fazem sentido
df.describe()

# parece que todos os valores estão corretos, e como não temos informações do contexto real dos aplicativos, não sabemos
# se a ordem de grandeza faz sentido, então vamos presumir que os dados estão na ordem de grandeza correta.
# Entretanto, os valores de "daily_ratings" e "daily_reviews" apresentam um valor mínimo muito menor que o esperado.
df.loc[
    (df["daily_ratings"] < -1000)
    | (df["daily_reviews"] < -1000)
]

# Olhando os dados, os dados não parecem ser discrepantes devido a grande quantidade de usuários envolvidos,
# o que torna mais possível que uma quantidade significativa de usuários retirem as suas avaliações

# Agora que todos os dados estão limpos, podemos armazenar o dataframe
pasta_atual = os.getcwd()
pasta_pai = os.path.dirname(pasta_atual)
pasta_data = os.path.join(pasta_pai, "data")

df.to_csv(f"{pasta_data}\\dados_limpos.csv", index=False)
