==> notebooks/3_extracao_de_caracteristicas.ipynb <==
## Extração de características

Esse notebook está responsável fazer a extraçãop de certas características dos dados, bem como preparar os dados para serem utilizados para treinar o modelo na etapa seguinte.

---

# importando pacotes
import os
import pandas as pd
import numpy as np

import datetime
import holidays

from sklearn.preprocessing import LabelEncoder

import pickle

# agora que temos os dados limpos, podemos seguir com a extralção de caracterpisticas e preparar parte dos dados
# para o treinamento do modelo
pasta_atual = os.getcwd()
pasta_pai = os.path.dirname(pasta_atual)
pasta_data = os.path.join(pasta_pai, "data")

df = pd.read_csv(f"{pasta_data}\\dados_limpos.csv", parse_dates=["date"])\
    .sort_values(by=["date", "appId"])\
    .reset_index(drop=True)

df

# Como todos os dados são referentes ao Brasil, vamos adicionar informações de feriados e finais de semana
# do Brasil, pois podem impactar o aesso de alguns aplicativos

feriados_brasil = holidays.Brazil()

def validar_se_dia_trabalho(data):
    ser_dia_semana = data.weekday() < 5
    ser_feriado = data in feriados_brasil

    return ser_dia_semana and not ser_feriado
df["dia_de_trabalho"] = df["date"].apply(validar_se_dia_trabalho)

# vamos também fazer a codificação ciclica das informações de tempo de:
    # semana
    # trimeste
    # ano
    # meia decada

def codificacao_ciclica(aux_df):
    df = aux_df.copy()

    timestamp = df["date"].map(pd.Timestamp.timestamp)

    dia = 24 * 60 * 60
    semana = dia * 7
    ano = 365.2425 * dia
    trimestre = ano / 4
    meia_decada = ano * 5

    df["sin_semana"] = np.sin(timestamp * (2 * np.pi / semana))
    df["cos_semana"] = np.cos(timestamp * (2 * np.pi / semana))

    df["sin_trimestre"] = np.sin(timestamp * (2 * np.pi / trimestre))
    df["cos_trimestre"] = np.cos(timestamp * (2 * np.pi / trimestre))

    df["sin_ano"] = np.sin(timestamp * (2 * np.pi / ano))
    df["cos_ano"] = np.cos(timestamp * (2 * np.pi / ano))

    df["sin_meia_decada"] = np.sin(timestamp * (2 * np.pi / meia_decada))
    df["cos_meia_decada"] = np.cos(timestamp * (2 * np.pi / meia_decada))

    return df

df = codificacao_ciclica(df)

# Também vamos adicionar dados moveis
def adicao_dados_moveis(aux_df, colunas, dias_janela):
    df = aux_df.copy()

    for coluna in colunas:
        df[f"{coluna}_media_{dias_janela}"] = df.groupby("appId")[coluna]\
            .rolling(window = dias_janela, min_periods = 1)\
            .mean()\
            .reset_index(level=0, drop=True)

        df[f"{coluna}_dp_{dias_janela}"] = df.groupby("appId")[coluna]\
            .rolling(window = dias_janela, min_periods = 1)\
            .std()\
            .reset_index(level=0, drop=True)\
            .fillna(0)

    return df

colunas_dados_moveis = ["dauReal", "mauReal", "newinstalls", "ratings", "reviews"] # esses dados são reduntantes com
# daily_ratings e daily_reviews, por isso não estão inclusos
# predictionLoss parecer ser fruto de um outro modelo e, como não tenho conhecimento de como esses dados são populados,
# vou evitar fazer muitas tratavias com esse dado para não complixar desnecessariamente o modelo

df = adicao_dados_moveis(df, colunas = colunas_dados_moveis, dias_janela = 3)
df = adicao_dados_moveis(df, colunas = colunas_dados_moveis, dias_janela = 7)


# Adicionar uma coluna para o dado anterior de dauReal
df["anterior_dauReal"] = df.groupby("appId")["dauReal"].shift(+1)

# Vamos também preparar uma coluna com o dado seguinte de "dauReal"
df["proximo_dauReal"] = df.groupby("appId")["dauReal"].shift(-1)

# converter a coluna de dados categoricos em dados dummy
df = pd.get_dummies(df, columns=['category'])

# Por fim, vamos fazer o LabelEncoding do appId para podermos utilizar o identificador do aplicativo
# como informação relevante para o modelo
le = LabelEncoder()
df["cod_appId"] = le.fit_transform(df["appId"])


df = df.dropna(subset=["proximo_dauReal", "anterior_dauReal"])

pasta_atual = os.getcwd()
pasta_pai = os.path.dirname(pasta_atual)
pasta_data = os.path.join(pasta_pai, "data")

df.to_csv(f"{pasta_data}\\dados_limpos_e_completos.csv", index=False)

with open(f"{pasta_data}\\label_encoder.pkl", 'wb') as f:
    pickle.dump(le, f)
